feat(tests): stop adding file system interface
The complexity of adding a fake filesystem interface is quite a bit more
work than I thought it would be at the outset. It must keep track of the
whole fake filesystem, in order to behave like a real one, and also to
be able to make assertions about its state.

An existing library does this, and provides easy integration with
pytest: pyfakefs.

I may decide to switch to using temporary directories in the tests, but
I think this will work well for now: it provides an easy way to plumb in
a fake filesystem, about which I can make assertions, without having to
make sure every function in a call chain is testable.